+++ src/components/ios/IOSNotch.vue
@@ -1,0 +1,18 @@
+<template>
+    <div class="notch">
+        
+    </div>
+</template>
+
+<script setup>
+    
+</script>
+
+<style scoped lang="scss">
+    .notch {
+        width: 200px;
+        height: 40px;
+        background-color: black;
+        border-radius: 100px;
+    }
+</style>
